Move tests to more appropriate place
